Refactor authentication redirection
