+++ apps/frontend/src/components/auth/RedirectIfAuth.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from '../hooks/useAuth';
+import { useAuth } from '../../hooks/useAuth';
 import { Navigate } from 'react-router-dom';
 
 export function RedirectIfAuth({ children }: { children: React.ReactNode }) {
